restructure shared files, implement sandbox verification
