+++ src/cli/tsconfig.json
@@ -7,7 +7,7 @@
     "moduleResolution": "node",
     "outDir": "../../dist/bin",
     "sourceMap": false,
-    "target": "es5",
+    "target": "es2015",
     "newLine": "LF",
     "typeRoots": [
       "../../node_modules/@types"
